mk3.c closed 19/19: the frame, the physics, and what 0x16462 really means
Tenth file complete in gamecode/logic. mk3.c is the fight loop, and it
settles several things the other twenty files could only circle.

THIRTY. The symbol gaps divide exactly by the four strides this project
had assembled one function at a time: GrObj 30x76, Plyr 30x108, Pp
30x140, mytc 30x268. mk3_init's wiring loop then links them -- Plyr[n]
points at Pp[n], mytc[n] and GrObj[n]. So MK3OBJ.field08, called
"another object" for twenty files, is the GrObj with the same index, and
the engine's hard limit is thirty live things at once.

THE PHYSICS IS FOUR WORDS. gravity_n_bounds adds 0x20 to 0x1c;
DisplayUpdate adds 0x18 to 0x0c and 0x1c to 0x10. x and y are 16.16
fixed point. A jump is one negative 0x1c and one positive 0x20.

0x16462 IS NOT A DURATION. Eleven sites recorded it as the
"park-and-never-wake duration" because fieldfc counts down. mk3_update
compares it against the RETURN VALUE and releases the thread on the
spot, clearing the owner's slave fields first. Corrected at all eleven
sites and documented as MK3_THREAD_DONE in mk3logic.h.

THE INPUT LAYER IS TranslateJoybits, and moves.c confirms it from the
other end: the bits it spreads land as exactly 0x00070070 and 0x00707000,
the two masks buttons_in_a2 was already picking between. Two independent
readings, agreeing to the bit. Keyboard and gamepad need to produce two
nine-bit words; there is no touch anywhere in the fight.

mk3_remap is VERIFIED, not argued. It touches no memory, so it was
emulated over all 131,072 inputs and the C was compiled and compared
against that table entry by entry. 69 animation numbers map to six.

Also: bit 7 of a character number is the CPU flag -- which GameCode.c
had already found from the call side, independently. The arena walls are
leftPlayerLimit+58 and rightPlayerLimit-48. G is 1144 bytes. A handler
returning 0 re-runs immediately; returning negative abandons the frame.

Corrected a hypothesis this session made and mk3_update disproved: bit
0x400 of an input word is a special-move request, not demo playback.

Seven prototype disagreements fixed. The eighth,
LIME_RenderMeshSingleIndexed, is pre-existing and NOT fixed: measurement
says the armv7 slice takes five core arguments plus an unread float, so
neither existing spelling is right. Recorded in Players.c and left for
protos.py to keep reporting.

gamecode/logic: 1545/2172 = 71.13%

+++ decomp/gamecode/logic/mkstat.c
@@ -1213,7 +1213,7 @@ long t_noogy_suspended(MK3THREAD *thread)
  *                       obj->field2c = f
  *                       if ((f & 4) != 0) {
  *                           jade_normpal(obj)
- *                           token := 0xf4, park 0x16462
+ *                           token := 0xf4, return 0x16462 (delete)
  *                       } else {
  *                           pop a level, or t_local_reaction_exit
  *                       }
@@ -1602,7 +1602,7 @@ long t_stat_do_duck_punch(MK3THREAD *thread)
  *
  *      token == 0x100:  obj->field1c = 0xc; ochar_sound(obj)
  *                       if (--obj->field40 > 0) -- back to the swap body --
- *                       token := 0x108, park 0x16462
+ *                       token := 0x108, return 0x16462 (delete)
  *
  *      otherwise:       return -3
  *
@@ -1947,7 +1947,7 @@ long t_stat_do_sweep_kick(MK3THREAD *thread)
  * the routine finishes in t_wait_forever -- the baby never does anything again.
  */
 extern uint32_t ochar_babies[];                   /* 0x00167350 */
-void mk3_getbbox(uint32_t ani, int *p1, int *p2, int *p3, int *p4);
+void mk3_getbbox(long ani, int *p1, int *p2, int *p3, int *p4);
 void tsound_func(MK3OBJ *obj, uint32_t arg);
 long t_wait_forever(MK3THREAD *thread);
